floats -> doubles, percep train&pred segfault :(

+++ funcs.c
@@ -3,7 +3,7 @@
 #include <omp.h>
 #include "matrix.h"
 
-float func_linear(float n) {
+double func_linear(double n) {
     return n;
 }
 
@@ -19,7 +19,7 @@ matrix *activation_linear(matrix *A) {
     return ans;
 }
 
-float func_relu(float n) {
+double func_relu(double n) {
     return n >= 0 ? n : 0;
 }
 
@@ -35,7 +35,7 @@ matrix *activation_relu(matrix *A) {
     return ans;
 }
 
-float func_sigmoid(float n) {
+double func_sigmoid(double n) {
     return 1 / (1 + expf(-n));
 }
 
@@ -51,7 +51,7 @@ matrix *activation_sigmoid(matrix *A) {
     return ans;
 }
 
-float func_tanh(float n) {
+double func_tanh(double n) {
     return (expf(n) - expf(-n)) / (expf(n) + expf(-n));
 }
 
@@ -67,7 +67,7 @@ matrix *activation_tanh(matrix *A) {
     return ans;
 }
 
-float derivative_func_linear(float n) {
+double derivative_func_linear(double n) {
     return 1;
 }
 
@@ -83,7 +83,7 @@ matrix *derivative_linear(matrix *A) {
     return ans;
 }
 
-float derivative_func_relu(float n) {
+double derivative_func_relu(double n) {
     if (n == 0) return NAN;
     else if (n > 0) return 1;
     else return 0;
@@ -101,7 +101,7 @@ matrix *derivative_relu(matrix *A) {
     return ans;
 }
 
-float derivative_func_sigmoid(float n) {
+double derivative_func_sigmoid(double n) {
     return func_sigmoid(n) * (1 - func_sigmoid(n));
 }
 
@@ -117,7 +117,7 @@ matrix *derivative_sigmoid(matrix *A) {
     return ans;
 }
 
-float derivative_func_tanh(float n) {
+double derivative_func_tanh(double n) {
     return 1 - pow(func_tanh(n), 2);
 }
